Devel

+++ skel-user/src/main/scala/io/syspulse/skel/user/server/UserRoutes.scala
@@ -1,10 +1,19 @@
 package io.syspulse.skel.user.server
 
+import com.typesafe.scalalogging.Logger
+import io.jvm.uuid._
+import scala.util.{Try,Success,Failure}
+import java.nio.file.Paths
+
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.server.Route
+import akka.stream.scaladsl.FileIO
 
 import scala.concurrent.Future
+import scala.concurrent.Await
+import scala.concurrent.duration.Duration
+
 import akka.actor.typed.ActorRef
 import akka.actor.typed.ActorSystem
 import akka.actor.typed.scaladsl.ActorContext
@@ -16,9 +25,9 @@ import akka.http.scaladsl.model.ContentTypes._
 import akka.http.scaladsl.model.headers.`Content-Type`
 import akka.http.scaladsl.server.RejectionHandler
 import akka.http.scaladsl.model.StatusCodes._
-import com.typesafe.scalalogging.Logger
 
-import io.jvm.uuid._
+import ch.megard.akka.http.cors.scaladsl.CorsDirectives._
+import ch.megard.akka.http.cors.scaladsl.settings.CorsSettings
 
 import io.swagger.v3.oas.annotations.enums.ParameterIn
 import io.swagger.v3.oas.annotations.media.{Content, Schema}
@@ -27,7 +36,7 @@ import io.swagger.v3.oas.annotations.{Operation, Parameter}
 import io.swagger.v3.oas.annotations.parameters.RequestBody
 // import javax.ws.rs.{Consumes, POST, GET, DELETE, Path, Produces}
 // import javax.ws.rs.core.MediaType
-import jakarta.ws.rs.{Consumes, POST, GET, DELETE, Path, Produces}
+import jakarta.ws.rs.{Consumes, POST, PUT, GET, DELETE, Path, Produces}
 import jakarta.ws.rs.core.MediaType
 
 
@@ -39,19 +48,16 @@ import io.syspulse.skel.service.CommonRoutes
 
 import io.syspulse.skel.Command
 
-import scala.concurrent.Await
-import scala.concurrent.duration.Duration
-
 import io.syspulse.skel.auth.permissions.rbac.Permissions
 import io.syspulse.skel.auth.RouteAuthorizers
 
 import io.syspulse.skel.user._
 import io.syspulse.skel.user.store.UserRegistry
 import io.syspulse.skel.user.store.UserRegistry._
+import io.syspulse.skel.user.server.{UserActionRes, Users, UserCreateReq, UserUpdateReq}
 
-
-@Path("/api/v1/user")
-class UserRoutes(registry: ActorRef[Command])(implicit context: ActorContext[_]) extends CommonRoutes with Routeable with RouteAuthorizers {
+@Path("/")
+class UserRoutes(registry: ActorRef[Command])(implicit context: ActorContext[_],config:Config) extends CommonRoutes with Routeable with RouteAuthorizers {
   //val log = Logger(s"${this}")
   implicit val system: ActorSystem[_] = context.system
   
@@ -65,12 +71,14 @@ class UserRoutes(registry: ActorRef[Command])(implicit context: ActorContext[_])
   val metricGetCount: Counter = Counter.build().name("skel_user_get_total").help("User gets").register(cr)
   val metricDeleteCount: Counter = Counter.build().name("skel_user_delete_total").help("User deletes").register(cr)
   val metricCreateCount: Counter = Counter.build().name("skel_user_create_total").help("User creates").register(cr)
+  val metricUpdateCount: Counter = Counter.build().name("skel_user_update_total").help("User updates").register(cr)
   
   def getUsers(): Future[Users] = registry.ask(GetUsers)
-  def getUser(id: UUID): Future[Option[User]] = registry.ask(GetUser(id, _))
+  def getUser(id: UUID): Future[Try[User]] = registry.ask(GetUser(id, _))
   def getUserByXid(xid: String): Future[Option[User]] = registry.ask(GetUserByXid(xid, _))
 
-  def createUser(userCreate: UserCreateReq): Future[Option[User]] = registry.ask(CreateUser(userCreate, _))
+  def createUser(req: UserCreateReq): Future[Try[User]] = registry.ask(CreateUser(req, _))
+  def updateUser(uid:UUID,req: UserUpdateReq): Future[Try[User]] = registry.ask(UpdateUser(uid,req, _))
   def deleteUser(id: UUID): Future[UserActionRes] = registry.ask(DeleteUser(id, _))
   def randomUser(): Future[User] = registry.ask(RandomUser(_))
 
@@ -142,16 +150,30 @@ class UserRoutes(registry: ActorRef[Command])(implicit context: ActorContext[_])
 
   @POST @Path("/") @Consumes(Array(MediaType.APPLICATION_JSON))
   @Produces(Array(MediaType.APPLICATION_JSON))
-  @Operation(tags = Array("user"),summary = "Create User Secret",
+  @Operation(tags = Array("user"),summary = "Create User",
     requestBody = new RequestBody(content = Array(new Content(schema = new Schema(implementation = classOf[UserCreateReq])))),
     responses = Array(new ApiResponse(responseCode = "200", description = "User",content = Array(new Content(schema = new Schema(implementation = classOf[User])))))
   )
   def createUserRoute = post {
-    entity(as[UserCreateReq]) { userCreate =>
-      onSuccess(createUser(userCreate)) { r =>
+    entity(as[UserCreateReq]) { req =>
+      onSuccess(createUser(req)) { r =>
         metricCreateCount.inc()
-        log.info(s"r = ${r}")
         complete(StatusCodes.Created, r)
+      }
+    }
+  }
+
+  @PUT @Path("/") @Consumes(Array(MediaType.APPLICATION_JSON))
+  @Produces(Array(MediaType.APPLICATION_JSON))
+  @Operation(tags = Array("user"),summary = "Update User",
+    requestBody = new RequestBody(content = Array(new Content(schema = new Schema(implementation = classOf[UserUpdateReq])))),
+    responses = Array(new ApiResponse(responseCode = "200", description = "User",content = Array(new Content(schema = new Schema(implementation = classOf[User])))))
+  )
+  def updateUserRoute(uid:String) = put {
+    entity(as[UserUpdateReq]) { req =>
+      onSuccess(updateUser(UUID(uid),req)) { r =>
+        metricUpdateCount.inc()
+        complete(StatusCodes.OK, r)
       }
     }
   }
@@ -163,66 +185,65 @@ class UserRoutes(registry: ActorRef[Command])(implicit context: ActorContext[_])
     }
   }
 
-  override def routes: Route =
+  def uploadFileRoute(uid:Option[UUID],fileField:String = "fileUpload",fileName:Option[String] = None,dir:String = config.uploadStore) = 
+    post {
+      fileUpload(fileField) {
+        case (fileInfo, fileStream) =>
+          val fileName0 = fileInfo.fileName
+          val ext = fileName0.split("\\.").lastOption.getOrElse("png")
+          val fileName1 = fileName.getOrElse(s"${uid.orElse(Some("")).get}-${UUID.random}.${ext}")
+          val fullName = s"${dir}/${fileName1}"
+          val sink = FileIO.toPath(Paths.get(dir) resolve fileName1)
+          val r = fileStream.runWith(sink)
+          onSuccess(r) { r =>
+            log.info(s"${r}: ${fileName0} -> ${fullName}: ${r.count}")
+            complete(OK,UserUploadRes("success",uid,config.uploadUri + "/" + fileName1,Some(fullName)))
+          }
+      }
+    }
+  
+  val corsAllow = CorsSettings(system.classicSystem)
+    //.withAllowGenericHttpRequests(true)
+    .withAllowCredentials(true)
+    .withAllowedMethods(Seq(HttpMethods.OPTIONS,HttpMethods.GET,HttpMethods.POST,HttpMethods.PUT,HttpMethods.DELETE,HttpMethods.HEAD))
+
+  override def routes: Route = cors(corsAllow) {
       concat(
         pathEndOrSingleSlash {
-          // authenticate()(authn =>
-          //   authorize(Permissions.isAdmin(authn)) {
-          //     concat(
-          //       getUsersRoute(),
-          //       createUserRoute
-          //     )
-          //   }
-          // )
           concat(
             authenticate()(authn =>
-              authorize(Permissions.isAdmin(authn)) {              
+              authorize(Permissions.isAdmin(authn) || Permissions.isService(authn)) {
                 getUsersRoute() ~                
                 createUserRoute  
               }
-            ),
-            //createUserRoute
+            ),            
           )
-        },
-        // pathPrefix("info") {
-        //   path(Segment) { userId => 
-        //     getUserInfo(userId)
-        //   }
-        // },
-        pathSuffix("random") {
-          createUserRandomRoute()
         },
         pathPrefix("xid") {
           pathPrefix(Segment) { id => 
-            getUserByXidRoute(id)
+            authenticate()(authn =>
+              getUserByXidRoute(id)
+            )
           }
         },
         pathPrefix(Segment) { id => 
-          // pathPrefix("eid") {
-          //   pathEndOrSingleSlash {
-          //     getUserByEidRoute(id)
-          //   } 
-          //   ~
-          //   path(Segment) { code =>
-          //     getUserCodeVerifyRoute(id,code)
-          //   }
-          // } ~
-
-          pathEndOrSingleSlash {
-            // concat(
-            //   getUserRoute(id),
-            //   deleteUserRoute(id),
-            // )          
+          pathPrefix("avatar") {
             authenticate()(authn =>
-              authorize(Permissions.isUser(UUID(id),authn)) {
-                getUserRoute(id)
+              uploadFileRoute(Some(UUID(id)))
+            )
+          } ~
+          pathEndOrSingleSlash {
+            authenticate()(authn =>
+              authorize(Permissions.isUser(UUID(id),authn) || Permissions.isAdmin(authn) || Permissions.isService(authn)) {
+                updateUserRoute(id) ~
+                getUserRoute(id)                 
               } ~
-              authorize(Permissions.isAdmin(authn)) {
+              authorize(Permissions.isAdmin(authn) || Permissions.isService(authn)) {
                 deleteUserRoute(id)
               }
             ) 
           }
         }
       )
-    
+  }
 }
